awafdg

+++ code/datums/wounds/pierce.dm
@@ -117,10 +117,16 @@
 	blood_flow -= 0.05 * power // 20u * 0.05 = -1 blood flow, less than with slashes but still good considering smaller bleed rates
 
 /datum/wound/pierce/on_silfrine(power)
-	if(severity < WOUND_SEVERITY_SEVERE)
-		qdel(src)
-	else
-		blood_flow -= 0.1 * power
+	switch(power)
+		if(0 to 5)
+			EMPTY_BLOCK_GUARD
+		if(6 to 11)
+			if(severity < WOUND_SEVERITY_MODERATE)
+				qdel(src)
+		if(12 to 30)
+			if(severity < WOUND_SEVERITY_SEVERE)
+				qdel(src)
+	blood_flow -= 0.05 * power'
 
 /// If someone is using a suture to close this puncture
 /datum/wound/pierce/proc/suture(obj/item/stack/medical/suture/I, mob/user)
